requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
pydantic==2.12.4
pydantic-settings==2.1.0
python-dotenv==1.2.1
SQLAlchemy==2.0.44
alembic==1.17.1
psycopg[binary]==3.2.12
redis==7.0.1
email-validator==2.2.0
